# Speed up file pins